refactoring: moved CCR to utils, moved useGetTroves to general hooks
