--- apps/frontend/src/app/2_molecules/SystemStats/utils.ts
@@ -1,5 +1,0 @@
-export const calculateCollateralRatio = (
-  collateral: string,
-  debt: string,
-  price: string,
-) => (((Number(collateral) * Number(price)) / Number(debt)) * 100).toFixed(2);
